Move message to appropriate location, hook up to provider/content-scripts
