+++ apps/extension/src/background/approvals/constants.ts
@@ -1,1 +1,1 @@
-export const ROUTE = "approvals";
+export const ROUTE = "approvals-route";
