+++ apps/extension/src/background/approvals/messages.ts
@@ -2,34 +2,8 @@ import { Message } from "router";
 import { ROUTE } from "./constants";
 
 enum MessageType {
-  ApproveTx = "approve-tx",
   RejectTx = "reject-tx",
   SubmitApprovedTx = "submit-approved-tx",
-}
-
-export class ApproveTxMsg extends Message<void> {
-  public static type(): MessageType {
-    return MessageType.ApproveTx;
-  }
-
-  constructor(public readonly txMsg: string) {
-    super();
-  }
-
-  validate(): void {
-    if (!this.txMsg || this.txMsg === "") {
-      throw new Error("txMsg was not provided!");
-    }
-    return;
-  }
-
-  route(): string {
-    return ROUTE;
-  }
-
-  type(): string {
-    return ApproveTxMsg.type();
-  }
 }
 
 export class RejectTxMsg extends Message<void> {
