+++ apps/extension/src/provider/Anoma.ts
@@ -2,6 +2,7 @@ import { Anoma as IAnoma, Chain, DerivedAccount } from "@anoma/types";
 import { Ports, MessageRequester } from "router";
 
 import {
+  ApproveTxMsg,
   ConnectInterfaceMsg,
   GetChainMsg,
   GetChainsMsg,
@@ -10,7 +11,6 @@ import {
   QueryAccountsMsg,
   SubmitBondMsg,
   SubmitUnbondMsg,
-  SubmitTransferMsg,
   SubmitIbcTransferMsg,
 } from "./messages";
 
@@ -74,7 +74,7 @@ export class Anoma implements IAnoma {
   public async submitTransfer(txMsg: string): Promise<void> {
     return await this.requester?.sendMessage(
       Ports.Background,
-      new SubmitTransferMsg(txMsg)
+      new ApproveTxMsg(txMsg)
     );
   }
